CORE-3203 Fixes for get-collaborators service.
Updated metadactyl/get-collaborators to filter out blank usernames before attempting to search them in the user_info/get-user-details service.
Updated user_info/get-user-details to look for usernames in the :users list in the map returned by search, and to return the successfully searched user info.
Updated user_info/empty-user-info to use a string for the :id field, since the search service returns strings for the IDs as well.