CORE-4140: Fixed sharing so that transitive sharing should now work.
The issue is that we were getting the home directory of the user doing
the sharing instead of the home directory of the user who has the file
in their home directory. This meant that transitive sharing wasn't
working because the other users to receive the file would be given
access to the second sharer's home directory and not the directory of
the user who actually had the file in their home directory.