CORE-5010 Added /filesystem/move-contents endpoint.
Added a filesystem.move/move-contents function that fetches all paths in a source directory and moves them to a destination directory.
Added a filesystem.validators/validate-num-paths-under-folder function, used by the new move-contents function, that will throw an ERR_TOO_MANY_PATHS exception if the total number of paths under a folder is more than the util.config/fs-max-paths-in-request value.