CORE-5012 Updated download-contents path cart.
Refactored donkey.services.filesystem.directory/get-paths-in-folder to allow it to accept a limit parameter (defaults to the max-paths config).
Updated donkey.services.filesystem.updown/do-download-contents to pass all paths in the given folder to the download function, using the count of paths in the folder as a limit for now.